+++ Tests/FluentAssertions.Equivalency.Specs/SelectionRulesSpecs.Accessibility.cs
@@ -13,13 +13,13 @@ public partial class SelectionRulesSpecs
         public void When_a_property_is_write_only_it_should_be_ignored()
         {
             // Arrange
-            var subject = new ClassWithWriteOnlyProperty
+            var expected = new ClassWithWriteOnlyProperty
             {
                 WriteOnlyProperty = 123,
                 SomeOtherProperty = "whatever"
             };
 
-            var expected = new
+            var subject = new
             {
                 SomeOtherProperty = "whatever"
             };